requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.0
djoser==2.2.0
psycopg2-binary==2.9.6
pillow==10.2.0
